Add initial support for linux and Mac DMG builds

+++ build/tasks/build-task.js
@@ -68,12 +68,16 @@ module.exports = function(grunt) {
     grunt.task.run('electron:macbuild');
 
     // If we're on Mac, go ahead and run appdmg
-    /*if (process.platform === 'darwin') {
+    if (process.platform === 'darwin') {
       if (fsp.existsSync(conf('appdmg.target.dest'))) {
         fs.rm(conf('appdmg.target.dest'));
       }
 
       grunt.task.run('appdmg');
-    }*/
+    }
+  });
+
+  grunt.registerTask('build-linux', 'Build the release application for Linux', function(){
+    grunt.task.run('electron:linbuild', 'electron-installer-debian', 'electron-installer-redhat');
   });
 };
